Rewrite to slash-commands
Problem: the bot doesn't reply in servers anymore, but does reply in
direct messages.

Solution: one can just enable the toggle to let the bot read message
contents in the developer portal, but slash commands are nice anyway,
so we've rewritten to those instead.
Just in case, in preparation for 100+ servers. Only 97 to go!